Export the landed public surface and document it against what actually runs
Phases 3, 5 and 6 had landed but were never wired into the package namespace, so
`from focusedgaze import GazeConfig` still failed and only `__version__` was
exported. The declared policy in __init__.py is that names appear as their phase
lands, because an __all__ entry that cannot be imported is worse than an honest
omission. That step had simply been missed. 27 names now export, and every one
was checked importable rather than assumed.

The module docstring carries the status split, since it is the first thing anyone
reads: implemented and tested, implemented but unproven, and not yet written.
Calibration sits in the middle group and says so, because a wrong polynomial does
not raise and silently plausible output is the failure mode most likely to be
mistaken for a calibration problem by the person using it.

Assets and the fitting internals are deliberately NOT lifted to the top level.
They stay reachable as focusedgaze.assets and focusedgaze.calibration. They are
tools for the CLI and for calibration time, not part of the everyday API, and a
curated surface is easier to keep stable than an exhaustive one.

docs/usage.md now covers configuration, results and status, the exception tree,
the asset registry and calibration, alongside the filter and positioning sections
it already had. The aspirational section is kept and still marked as not running.

Every example in the "works today" section was EXECUTED, not eyeballed. That
caught one: the registry is a dict keyed by asset name, not a sequence, so the
documented loop would have raised AttributeError on a str for anyone who copied
it. A documentation example that has not been run is a claim, not a demonstration,
which is the same standard section 37.10 arrived at for the filter demo.

Also fixes the lint and typing debt the interrupted agents left, so the Phase 3
exit criterion holds again: ruff clean, mypy --strict clean over 25 files. The
one substantive fix is _terms, whose annotation admitted only arrays while apply
deliberately calls it with scalars, evaluating one reading at a time rather than
routing through the batch path so that a one-ULP difference cannot change which
samples the robust fitter rejects. The annotation was wrong, not the code, and it
now says so in the docstring.

194 tests still pass. Note that this changes no numerics.

+++ src/focusedgaze/calibration/fitter.py
@@ -267,7 +267,7 @@ def _holdout_error(
     n = int(data.shape[0])
     if fraction <= 0.0:
         return None
-    n_val = int(round(n * fraction))
+    n_val = round(n * fraction)          # round() with one argument already returns int
     n_train = n - n_val
     n_terms = int(polynomial_powers(degree).shape[0])
     if n_val < 1 or n_train < n_terms:
